+++ packages/components/src/components/ExampleList.styles.tsx
@@ -1,0 +1,5 @@
+import tw, {css, styled} from 'twin.macro';
+
+export const Examples = styled.div(() => [
+    tw`flex flex-wrap lg:flex-nowrap gap-8`,
+]);
