+++ packages/components/src/components/MarketplaceList.tsx
@@ -10,111 +10,18 @@ import {
   Table,
   TableBody,
   TableHeader,
-  TableItem,
-  TableItemButton,
-  TableItemDate,
-  TableItemImage,
-  TableItemInfo,
   TablePagination,
   Wrapper,
 } from './MarketplaceList.styles';
 
 import {
   IMarketplaceListProps,
-  IMarketplaceItemsProps,
   IMarketplaceItemProps,
 } from '../types/components';
 import { useThemeContext } from '../helpers/theme';
 import { marketplaceThemedAssets } from '../helpers/assets';
 import { toggleLockBodyScroll } from '../helpers/modals';
-import { Tag, TagsContainer } from './Tag';
-
-const TableItems: React.FC<IMarketplaceItemsProps> = ({
-  icon,
-  items,
-  handleModal,
-  setCurrentItem,
-  ...restProps
-}) => {
-  const formatDate = (value: string) => {
-    const months = [
-      'Jan',
-      'Feb',
-      'Mar',
-      'Apr',
-      'May',
-      'Jun',
-      'Jul',
-      'Aug',
-      'Sep',
-      'Oct',
-      'Nov',
-      'Dec',
-    ];
-    const date = new Date(value);
-    return `${
-      months[date.getMonth()]
-    } ${date.getDate()}, ${date.getFullYear()}`;
-  };
-
-  return (
-    <>
-      {items &&
-        items.map((item) => (
-          <TableItem key={item.title}>
-            <td>
-              {item.image && (
-                <TableItemImage {...item.image} {...restProps.imageProps} />
-              )}
-            </td>
-            <td>
-              <TableItemInfo>
-                <a
-                  {...(item.link
-                    ? item.link
-                    : {
-                        onClick: () => {
-                          setCurrentItem(item);
-                          handleModal(true);
-                        },
-                      })}
-                  {...restProps.linkProps}
-                >
-                  <span>
-                    <h3 {...restProps.titleProps}>{item.title}</h3>
-                    <p {...restProps.descriptionProps}>{item.description}</p>
-                    {item.tags && item.tags.length > 0 ? (
-                      <TagsContainer>
-                        {item.tags.map((tagName) => (
-                          <Tag key={tagName}>{tagName}</Tag>
-                        ))}
-                      </TagsContainer>
-                    ) : null}
-                  </span>
-                </a>
-              </TableItemInfo>
-            </td>
-            <td>
-              <TableItemDate {...restProps.dateProps}>
-                {formatDate(item.update)}
-              </TableItemDate>
-            </td>
-            <td>
-              <TableItemButton
-                type="button"
-                onClick={() => {
-                  handleModal(true);
-                  setCurrentItem(item);
-                }}
-              >
-                <img src={icon} alt=">" />
-              </TableItemButton>
-            </td>
-          </TableItem>
-        ))}
-    </>
-  );
-};
+import TableItems from './TableItems';
 
 export const MarketplaceList: React.FC<IMarketplaceListProps> = ({
   title,
